Rearranged triggers tests to better match the master branch
